# PM-BE/pm-application/src/main/resources/swagger/api.yml
...
openapi: '3.0.1'
info:
  title: 'JobsGo API'
  description: JobsGo API documentation
  contact: {}
  version: 0.0.1
servers:
  - url: http://localhost:8080/api/v1
    description: Development server
  - url: https://localhost:8080/api/v1
    description: Development server with TLS Profile
...
paths:
...
  /api/v1/profile/projects:
    get:
      tags:
        - project-controller
      summary: getProjects
      operationId: getProjectsUsingGET
      responses:
        '200':
          description: OK
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/CollectionModelOfEntityModelOfProject'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
    post:
      tags:
        - project-controller
      summary: createProject
      operationId: createProjectUsingPOST
      requestBody:
        content:
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/ProjectDTO'
      responses:
        '200':
          description: OK
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/EntityModelOfProject'
        '201':
          description: Created
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
  '/api/v1/profile/projects/{projectId}':
    get:
      tags:
        - project-controller
      summary: getProject
      operationId: getProjectUsingGET
      parameters:
        - name: projectId
          in: path
          description: projectId
          required: true
          style: simple
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/EntityModelOfProject'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
    delete:
      tags:
        - project-controller
      summary: deleteProject
      operationId: deleteProjectUsingDELETE
      parameters:
        - name: projectId
          in: path
          description: projectId
          required: true
          style: simple
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
        '204':
          description: No Content
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
    patch:
      tags:
        - project-controller
      summary: updateProject
      operationId: updateProjectUsingPATCH
      parameters:
        - name: projectId
          in: path
          description: projectId
          required: true
          style: simple
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/ProjectDTO'
      responses:
        '200':
          description: OK
          content:
            application/hal+json:
              schema:
                $ref: '#/components/schemas/EntityModelOfProject'
        '204':
          description: No Content
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
...
components:
...
  securitySchemes:
    oauth:
      type: oauth2
      description: OAuth2 authentication with KeyCloak
      flows:
        authorizationCode:
          authorizationUrl: http://localhost:8090/auth/realms/personal-management/protocol/openid-connect/auth
          tokenUrl: http://localhost:8090/auth/realms/personal-management/protocol/openid-connect/token
          scopes:
            jhipster: Jhipster specific claims
            email: Email claims
            profile: Profile claims
    openId:
      type: openIdConnect
      description: OpenID Connect authentication with KeyCloak
      openIdConnectUrl: http://localhost:8090/auth/realms/personal-management/.well-known/openid-configuration
  schemas:
...
    Author:
      title: Author
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        publication:
          $ref: '#/components/schemas/Publication'
        publicationId:
          type: integer
          format: int64
    Certification:
      title: Certification
      type: object
      properties:
        credentialID:
          type: string
        credentialURL:
          type: string
        expDate:
          type: string
          format: date-time
        id:
          type: integer
          format: int64
        issDate:
          type: string
          format: date-time
        issOrganization:
          type: string
        name:
          type: string
        personalProfile:
          $ref: '#/components/schemas/PersonalProfile'
        personalProfileId:
          type: integer
          format: int64
...
    CollectionModelOfEntityModelOfProject:
      title: CollectionModelOfEntityModelOfProject
      type: object
      properties:
        _embedded:
          type: array
          items:
            $ref: '#/components/schemas/EntityModelOfProject'
        _links:
          $ref: '#/components/schemas/Links'
...
    Creator:
      title: Creator
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        project:
          $ref: '#/components/schemas/Project'
        projectId:
          type: integer
          format: int64
    Education:
      title: Education
      type: object
      properties:
        activities:
          type: string
        degree:
          type: string
        description:
          type: string
        endDate:
          type: string
          format: date-time
        fieldOfStudy:
          type: string
        grade:
          type: string
        id:
          type: integer
          format: int64
        personalProfile:
          $ref: '#/components/schemas/PersonalProfile'
        personalProfileId:
          type: integer
          format: int64
        school:
          type: string
        startDate:
          type: string
          format: date-time
...
    EntityModelOfProject:
      title: EntityModelOfProject
      type: object
      properties:
        _links:
          $ref: '#/components/schemas/Links'
        creators:
          uniqueItems: true
          type: array
          items:
            $ref: '#/components/schemas/Creator'
        description:
          type: string
        endDate:
          type: string
          format: date-time
        id:
          type: integer
          format: int64
        name:
          type: string
        personalProfile:
          $ref: '#/components/schemas/PersonalProfile'
        personalProfileId:
          type: integer
          format: int64
        startDate:
          type: string
          format: date-time
        url:
          type: string
...
    Links:
      title: Links
      type: object
      properties:
        empty:
          type: boolean
    PersonalProfile:
      title: PersonalProfile
      type: object
      properties:
        about:
          type: string
        address:
          type: string
        bgImageUrl:
          type: string
        birthday:
          type: string
          format: date-time
        certifications:
          uniqueItems: true
          type: array
          items:
            $ref: '#/components/schemas/Certification'
        country:
          type: string
        educations:
          uniqueItems: true
          type: array
          items:
            $ref: '#/components/schemas/Education'
        headline:
          type: string
        id:
          type: integer
          format: int64
        industry:
          type: string
        location:
          type: string
        phoneNumber:
          pattern: '^\+(?:[0-9]){6,14}[0-9]$'
          type: string
        projects:
          uniqueItems: true
          type: array
          items:
            $ref: '#/components/schemas/Project'
        publications:
          uniqueItems: true
          type: array
          items:
            $ref: '#/components/schemas/Publication'
        skills:
          uniqueItems: true
          type: array
          items:
            $ref: '#/components/schemas/Skill'
        user:
          $ref: '#/components/schemas/User'
        userId:
          type: string
        workExperiences:
          uniqueItems: true
          type: array
          items:
            $ref: '#/components/schemas/WorkExperience'
    Project:
      title: Project
      type: object
      properties:
        creators:
          uniqueItems: true
          type: array
          items:
            $ref: '#/components/schemas/Creator'
        description:
          type: string
        endDate:
          type: string
          format: date-time
        id:
          type: integer
          format: int64
        name:
          type: string
        personalProfile:
          $ref: '#/components/schemas/PersonalProfile'
        personalProfileId:
          type: integer
          format: int64
        startDate:
          type: string
          format: date-time
        url:
          type: string
    Publication:
      title: Publication
      type: object
      properties:
        authors:
          uniqueItems: true
          type: array
          items:
            $ref: '#/components/schemas/Author'
        description:
          type: string
        id:
          type: integer
          format: int64
        personalProfile:
          $ref: '#/components/schemas/PersonalProfile'
        personalProfileId:
          type: integer
          format: int64
        publicationDate:
          type: string
          format: date-time
        publisher:
          type: string
        title:
          type: string
        url:
          type: string
    Skill:
      title: Skill
      type: object
      properties:
        category:
          type: string
        id:
          type: integer
          format: int64
        name:
          type: string
        personalProfile:
          $ref: '#/components/schemas/PersonalProfile'
        personalProfileId:
          type: integer
          format: int64
    User:
      title: User
      required:
        - activated
        - login
      type: object
      properties:
        activated:
          type: boolean
        email:
          maxLength: 254
          minLength: 5
          type: string
        firstName:
          maxLength: 50
          minLength: 0
          type: string
        id:
          type: string
        imageUrl:
          maxLength: 256
          minLength: 0
          type: string
        langKey:
          maxLength: 10
          minLength: 2
          type: string
        lastName:
          maxLength: 50
          minLength: 0
          type: string
        login:
          maxLength: 50
          minLength: 1
          pattern: '^(?>[a-zA-Z0-9!$&*+=?^_`{|}~.-]+@[a-zA-Z0-9-]+(?:\.[a-zA-Z0-9-]+)*)|(?>[_.@A-Za-z0-9-]+)$'
          type: string
...
    WorkExperience:
      title: WorkExperience
      type: object
      properties:
        company:
          type: string
        employmentType:
          type: string
          enum:
            - FREE_LANCE
            - FULL_TIME
            - INTERNSHIP
            - PART_TIME
            - SELF_EMPLOYED
        endDate:
          type: string
          format: date-time
        id:
          type: integer
          format: int64
        location:
          type: string
        personalProfile:
          $ref: '#/components/schemas/PersonalProfile'
        personalProfileId:
          type: integer
          format: int64
        startDate:
          type: string
          format: date-time
        title:
          type: string
...
    ProjectDTO:
      title: ProjectDTO
      type: object
      required:
        - name
      properties:
        name:
          type: string
        startMonth:
          type: string
          pattern: '^0[1-9]$|^1[0-2]$'
        startYear:
          type: string
        endMonth:
          type: string
          pattern: '^0[1-9]$|^1[0-2]$'
        endYear:
          type: string
        creators:
          type: string
        description:
          type: string
        url:
          type: string
...
security:
  - oauth: [jhipster, email, profile]
  - openId: [jhipster, email, profile]